Changed free request to 6 a day
The reason for this change is for frontend tests to pass easier. 6 tests run in total between browsers for book generation on same email, so to accomodate that, I made this change. Having said that, it would probably be better to always use a unique email for that tests to not even run into this problem, but oh well.